requirements.txt
numpy==2.2.4
opencv-python==4.11.0.86
pillow==11.1.0
pycryptodome==3.22.0
